review#0: applied modularization
