+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>이벤트 - TODOS</title>
     <link rel="stylesheet" href="css/style.css" />
-    <script src="application.js"></script>
+    <script src="application.js" type="module"></script>
   </head>
   <body>
     <div class="todoapp">
